myshop/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.1.8
django-sslserver==0.22
pillow==10.3.0